backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==2.2.19
django-extra-fields==3.0.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
install==1.3.5
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
Pillow==9.1.0
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
